requirements.txt
psycopg2-binary==2.9.9
prometheus_client==0.19.0
python-dotenv==1.0.0